+++ src/builders/fromNodePackageJson.node.ts
@@ -30,10 +30,10 @@ import type { PackageJson } from '../_helpers/packageJson'
 import { splitNameGroup } from '../_helpers/packageJson'
 import { ComponentType } from '../enums/componentType'
 import type * as Factories from '../factories/index.node'
-import { Tool } from '../models/tool'
-import { ExternalReferenceRepository } from '../models/externalReference'
 import { Component } from '../models/component'
+import { ExternalReferenceRepository } from '../models/externalReference'
 import { LicenseRepository } from '../models/license'
+import { Tool } from '../models/tool'
 
 /**
  * Node-specific ToolBuilder.
