+++ src/factories/packageUrl.ts
@@ -21,7 +21,7 @@ import { PackageURL } from 'packageurl-js'
 
 import { PackageUrlQualifierNames } from '../_helpers/packageUrl'
 import { ExternalReferenceType } from '../enums/externalReferenceType'
-import type { Component } from '../models'
+import type { Component } from '../models/component'
 
 export class PackageUrlFactory {
   readonly #type: PackageURL['type']
