+++ src/serialize/json/normalize.ts
@@ -23,12 +23,12 @@ import type { Stringable } from '../../_helpers/stringable'
 import { treeIteratorSymbol } from '../../_helpers/tree'
 import { escapeUri } from '../../_helpers/uri'
 import type * as Models from '../../models'
+import { LicenseExpression, NamedLicense, SpdxLicense } from '../../models/license'
 import { NamedLifecycle } from '../../models/lifecycle'
 import { AffectedSingleVersion, AffectedVersionRange } from '../../models/vulnerability/affect'
-import { NamedLicense, SpdxLicense, LicenseExpression } from '../../models/license'
 import { isSupportedSpdxId } from '../../spdx'
-import { Version as SpecVersion } from '../../spec'
 import type { _SpecProtocol as Spec } from '../../spec/_protocol'
+import { Version as SpecVersion } from '../../spec/enums'
 import type { NormalizerOptions } from '../types'
 import type { Normalized } from './types'
 import { JsonSchema } from './types'
