+++ src/serialize/xml/normalize.ts
@@ -23,12 +23,12 @@ import type { Stringable } from '../../_helpers/stringable'
 import { treeIteratorSymbol } from '../../_helpers/tree'
 import { escapeUri } from '../../_helpers/uri'
 import type * as Models from '../../models'
-import { NamedLifecycle} from '../../models/lifecycle'
-import { NamedLicense, SpdxLicense, LicenseExpression } from '../../models/license'
+import { LicenseExpression, NamedLicense, SpdxLicense } from '../../models/license'
+import { NamedLifecycle } from '../../models/lifecycle'
 import { AffectedSingleVersion, AffectedVersionRange } from '../../models/vulnerability/affect'
 import { isSupportedSpdxId } from '../../spdx'
-import { Version as SpecVersion } from '../../spec/enums'
 import type { _SpecProtocol as Spec } from '../../spec/_protocol'
+import { Version as SpecVersion } from '../../spec/enums'
 import type { NormalizerOptions } from '../types'
 import type { SimpleXml } from './types'
 import { XmlSchema } from './types'
